+++ src/engine/engine2/constants.ts
@@ -1,1 +1,6 @@
 export const CARD_INTERNAL_COOLDOWN_FIRES_PER_SECOND = 5;
+/**
+ * Tick rate in milliseconds
+ */
+
+export const TICK_RATE = 100;
