+++ src/engine/engine2/engine2Adapter.ts
@@ -2,13 +2,13 @@ import {
   Engine2,
   getCardAttribute as getCardAttribute2,
   BoardCardID,
-} from "./engine2/engine2";
-import { GameState as Engine2GameState } from "./engine2/engine2";
-import { getInitialGameState } from "./GameState";
-import { MonsterConfig, PlayerConfig } from "./GameState";
+} from "./engine2";
+import { GameState as Engine2GameState } from "./engine2";
+import { getInitialGameState } from "../GameState";
+import { MonsterConfig, PlayerConfig } from "../GameState";
 import { Cards, AttributeType } from "@/types/cardTypes";
 import { EncounterDays } from "@/types/encounterTypes";
-import { getTooltips as originalGetTooltips } from "./Engine";
+import { getTooltips as originalGetTooltips } from "../Engine";
 
 /**
  * Create a game state compatible with Engine2 from the parameters used by original getInitialGameState
@@ -111,19 +111,4 @@ export function runWithConfig(
 
   // Run simulation
   return run(initialState, maxTicks);
-}
-
-/**
- * Adapter for the getTooltips function to work with Engine2
- *
- * This function converts the Engine2 parameters to the format expected by the original getTooltips function,
- * allowing tooltip functionality to work with the Engine2 implementation.
- */
-export function getTooltips(
-  gameState: Engine2GameState,
-  playerID: number,
-  boardCardID: number,
-): string[] {
-  // Use the original implementation but with the Engine2 game state
-  return originalGetTooltips(gameState, playerID, boardCardID);
 }
